+++ includes/PF_FormLinker.php
@@ -70,7 +70,7 @@ class PFFormLinker {
 			$wgPageFormsFormPrinter->formHTML(
 				$formDefinition, false, false, null, $preloadContent,
 				'Some very long page name that will hopefully never get created ABCDEF123',
-				null, false, false, true, $inQueryArr
+				null, false, false, true, $inQueryArr, null, RequestContext::getMain()->getRequest()
 			);
 		$params = [];