+++ includes/PF_Hooks.php
@@ -255,7 +255,8 @@ class PFHooks {
 		$form_definition = StringUtils::delimiterReplace( '<noinclude>', '</noinclude>', '', $editpage->textbox1 );
 		[ $form_text, $data_text, $form_page_title, $generated_page_name, $parserOutput ] =
 			$wgPageFormsFormPrinter->formHTML(
-				$form_definition, null, false, null, null, "Page Forms form preview dummy title", null
+				$form_definition, null, false, null, null, "Page Forms form preview dummy title", null,
+				false, false, false, [], null, $request
 			);
 
 		$wgOut->addParserOutputMetadata( $parserOutput );